refactor(Calculator -> NumberAdder): 책임 분배상 클래스 네이밍 변경
