+++ src/test/java/calculator/domain/NumberAdderTest.java
@@ -2,18 +2,19 @@ package calculator.domain;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import calculator.domain.adder.NumberAdder;
 import calculator.domain.parser.InputParser.ParsedInput;
 import org.junit.jupiter.api.Test;
 
-public class CalculatorTest {
+public class NumberAdderTest {
 
-    private final Calculator calculator = new Calculator();
+    private final NumberAdder numberAdder = new NumberAdder();
 
     @Test
     void 숫자_하나_테스트() {
         ParsedInput parsedInput = new ParsedInput(new String[]{"1"});
 
-        String result = calculator.adder(parsedInput);
+        String result = numberAdder.adder(parsedInput);
 
         assertEquals("1", result);
     }
@@ -22,7 +23,7 @@ public class CalculatorTest {
     void 숫자_두개_테스트() {
         ParsedInput parsedInput = new ParsedInput(new String[]{"1", "2"});
 
-        String result = calculator.adder(parsedInput);
+        String result = numberAdder.adder(parsedInput);
 
         assertEquals("3", result);
     }
@@ -31,7 +32,7 @@ public class CalculatorTest {
     void 숫자_세개_테스트() {
         ParsedInput parsedInput = new ParsedInput(new String[]{"1", "2", "3"});
 
-        String result = calculator.adder(parsedInput);
+        String result = numberAdder.adder(parsedInput);
 
         assertEquals("6", result);
     }
@@ -40,7 +41,7 @@ public class CalculatorTest {
     void int_범위_초과_테스트() {
         ParsedInput parsedInput = new ParsedInput(new String[]{"1000000000000", "20", "3"});
 
-        String result = calculator.adder(parsedInput);
+        String result = numberAdder.adder(parsedInput);
 
         assertEquals("1000000000023", result);
     }
@@ -49,7 +50,7 @@ public class CalculatorTest {
     void 소수점_테스트() {
         ParsedInput parsedInput = new ParsedInput(new String[]{"1.1", "2.2", "3.3"});
 
-        String result = calculator.adder(parsedInput);
+        String result = numberAdder.adder(parsedInput);
 
         assertEquals("6.6", result);
     }
